+++ Genius/Models/Localization/VaryingLocalizationTable.swift
@@ -8,15 +8,17 @@ import Foundation
 protocol VaryingLocalizationTable {
 
 	static var administratorPrivilegesRequest: LocalizedStringResource { get }
-	static var askToAllowAccessoriesToConnect: LocalizedStringResource { get }
-	static var automaticUpdates: LocalizedStringResource { get }
-	static var automaticAppUpdates: LocalizedStringResource { get }
 	static var appUpdates: LocalizedStringResource { get }
+	static var askToAllowAccessoriesToConnect: LocalizedStringResource { get }
+	static var automaticAppUpdates: LocalizedStringResource { get }
+	static var automaticUpdates: LocalizedStringResource { get }
+	static var board: LocalizedStringResource { get }
 	static var bootMode: LocalizedStringResource { get }
 	static var bootVolume: LocalizedStringResource { get }
 	static var checkForMacOSUpdates: LocalizedStringResource { get }
 	static var checkForUpdatesEllipsis: LocalizedStringResource { get }
 	static var cpuFrequency: LocalizedStringResource { get }
+	static var cpuIdentifier: LocalizedStringResource { get }
 	static var cpuName: LocalizedStringResource { get }
 	static var cpuType: LocalizedStringResource { get }
 	static var downloadMacOSUpdates: LocalizedStringResource { get }