+++ Genius/Models/Maintenance/MaintenanceCheck.swift
@@ -3,6 +3,7 @@
 // See LICENSE.txt for license information.
 //
 
+import _Concurrency
 import Defaults
 import SFSafeSymbols
 import SwiftUI
@@ -16,11 +17,11 @@ struct MaintenanceCheck<
 
 	let valueWrapper: ValueWrapper
 	let requirement: Wrapped
-	let applicable: Bool?
+	let available: Bool?
 
 	var uiRepresentation: Symbol? { get async {
-		if !?applicable ?? false {
-			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notApplicable) : nil
+		if !?available ?? false {
+			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notAvailable) : nil
 		} else if let value = await value.optional {
 			value >= requirement ? // swiftlint:disable:this void_function_in_ternary
 				Symbol(.checkmark, color: .green, label: .enabled) :
@@ -34,20 +35,33 @@ struct MaintenanceCheck<
 
 extension MaintenanceCheck where ValueWrapper == SyncValueWrapper<Value> {
 
+	@MainActor var uiRepresentation: Symbol? {
+		if !?available ?? false {
+			Defaults[.developmentMode] ? Symbol(.minus, color: .primary, label: .notAvailable) : nil
+		} else if let value = value.optional {
+			value >= requirement ? // swiftlint:disable:this void_function_in_ternary
+				Symbol(.checkmark, color: .green, label: .enabled) :
+				Symbol(.xmark, color: .red, label: .disabled)
+		} else {
+			Defaults[.developmentMode] || Defaults[.interfaceMode] >= .advanced ?
+				Symbol(.questionmark, color: .red, label: .unknown) : nil
+		}
+	}
+
 	// periphery:ignore
 	init(_ value: Value, requirement: Wrapped = .max) {
 		self.valueWrapper = SyncValueWrapper(wrappedValue: value)
 		self.requirement = requirement
-		self.applicable = true
+		self.available = true
 	}
 
 	init<Wrapped>(
 		_ value: @autoclosure () -> Value,
 		requirement: Wrapped = .max,
-		applicable: Bool?,
+		available: Bool?,
 	) where Value == Wrapped? {
-		self.valueWrapper = SyncValueWrapper(wrappedValue: applicable ?? true ? value() : nil)
-		self.applicable = applicable
+		self.valueWrapper = SyncValueWrapper(wrappedValue: available ?? true ? value() : nil)
+		self.available = available
 		self.requirement = requirement
 	}
 }
@@ -58,18 +72,18 @@ extension MaintenanceCheck where ValueWrapper == AsyncValueWrapper<Value> {
 	init(_ valueClosure: @escaping @Sendable () async -> Value, requirement: Wrapped = .max) {
 		self.valueWrapper = AsyncValueWrapper(valueClosure: valueClosure)
 		self.requirement = requirement
-		self.applicable = true
+		self.available = true
 	}
 
 	init<Wrapped>(
 		_ valueClosure: @escaping @Sendable () async -> Value,
 		requirement: Wrapped = .max,
-		applicable: Bool?,
+		available: Bool?,
 	) where Value == Wrapped? {
 		self.valueWrapper = AsyncValueWrapper(
-			valueClosure: applicable ?? true ? valueClosure : { @Sendable in nil },
+			valueClosure: available ?? true ? valueClosure : { @Sendable in nil },
 		)
 		self.requirement = requirement
-		self.applicable = applicable
+		self.available = available
 	}
 }