+++ Genius/Models/SystemInformation/SystemInformation.Software.swift
@@ -13,7 +13,7 @@ extension SystemInformation {
 
 			static let version = SystemInformationData<String?, _>(
 				{ await SystemProfiler.hardware?["SMC_version_system"] as? String },
-				applicable: Hardware.securityChip.value <=? .t1 &&? SystemProfiler.available,
+				available: Hardware.securityChip.value <=? .t1 &&? SystemProfiler.available,
 			)
 		}
 
@@ -21,7 +21,7 @@ extension SystemInformation {
 
 			static let version = SystemInformationData<String?, _>(
 				{ await SystemProfiler.hardware?["boot_rom_version"] as? String },
-				applicable: SystemProfiler.available,
+				available: SystemProfiler.available,
 			)
 		}
 
@@ -38,19 +38,21 @@ extension SystemInformation {
 			static let version = SystemInformationData<VersionNumber?, _>(
 				Sysctl.read("kern.osproductversion").flatMap(VersionNumber.init),
 			)
-			static let codeName = SystemInformationData<String?, _>({
-				if #unavailable(macOS 14) {
-					"Ventura"
-				} else if #unavailable(macOS 15) {
-					"Sonoma"
-				} else if #unavailable(macOS 16) {
-					"Sequoia"
-				} else if #unavailable(macOS 26) {
-					nil
-				} else if #unavailable(macOS 27) {
-					"Tahoe"
-				} else { nil }
-			}())
+			static let codeName = SystemInformationData<String?, _>(
+				PrivateFramework.aboutSettings.ASPlatformInfo?.shared?.osTitleString ?? {
+					if #unavailable(macOS 14) {
+						"macOS Ventura"
+					} else if #unavailable(macOS 15) {
+						"macOS Sonoma"
+					} else if #unavailable(macOS 16) {
+						"macOS Sequoia"
+					} else if #unavailable(macOS 26) {
+						nil
+					} else if #unavailable(macOS 27) {
+						"macOS Tahoe"
+					} else { nil }
+				}(),
+			)
 			static let buildNumber = SystemInformationData<String?, _>(Sysctl.read("kern.osversion"))
 			static let bootMode = SystemInformationData<BootMode?, _>({
 				if FileManager.default.fileExists(atPath: "/System/Library/CoreServices/Finder.app") {
@@ -61,19 +63,17 @@ extension SystemInformation {
 			}())
 			static let bootVolume = SystemInformationData<String?, _>(
 				{ await SystemProfiler.software?["boot_volume"] as? String },
-				applicable: SystemProfiler.available,
+				available: SystemProfiler.available,
 			)
 			static let loaderVersion = SystemInformationData<String?, _>(
 				{ await SystemProfiler.hardware?["os_loader_version"] as? String },
-				applicable: SystemProfiler.available,
+				available: SystemProfiler.available,
 			)
 		}
 
 		enum Computer {
 
-			static let name = SystemInformationData<String?, _>(
-				{ await SystemProfiler.software?["local_host_name"] as? String ?? Host.current().localizedName },
-			)
+			static let name = SystemInformationData<String?, _>(Host.current().localizedName)
 			static let hostName = SystemInformationData<String?, _>(Sysctl.read("kern.hostname"))
 		}