+++ Genius/Views/SystemInformation/SystemInformationHardwareView.swift
@@ -12,13 +12,16 @@ struct SystemInformationHardwareView: View {
 			.model: [
 				.name: SystemInformation.Hardware.Model.displayName,
 				.identifier: SystemInformation.Hardware.Model.identifier,
+				.board: SystemInformation.Hardware.Model.board,
 				.number: SystemInformation.Hardware.Model.number,
+				.configurationCode: SystemInformation.Hardware.Model.configCode,
 				.regulatoryNumber: SystemInformation.Hardware.Model.regulatoryNumber,
 			],
 			.specifications: [
 				.securityChip: SystemInformation.Hardware.securityChip,
 				.cpuType: SystemInformation.Hardware.CPU.type,
 				.cpuName: SystemInformation.Hardware.CPU.name,
+				.cpuIdentifier: SystemInformation.Hardware.CPU.identifier,
 				.cpuFrequency: SystemInformation.Hardware.CPU.frequency,
 				SystemInformation.Hardware.CPU.Cores.differentTypes ?? true ? .totalCores : .cores:
 					SystemInformation.Hardware.CPU.Cores.total,