+++ CHANGELOG.md
@@ -1,10 +1,15 @@
+<!--
+© 2025 F1248 <f1248@mailbox.org>
+See LICENSE.txt for license information.
+-->
+
 # Changelog
 
 ## Unreleased
 
 ### Added
 
-### Changed
+### Improved
 
 ### Fixed
 
@@ -12,7 +17,39 @@
 
 ### Dependencies
 
-### [Commit History](https://github.com/F1248/Genius/compare/v0.3.0...HEAD)
+### [Commit History](https://github.com/F1248/Genius/compare/v0.4.0...HEAD)
+
+## [0.4.0](https://github.com/F1248/Genius/releases/tag/v0.4.0)
+
+### Added
+
+- Make Model → Configuration Code and Name available on devices with 10-digit serial numbers
+- Add Background Security Improvements introduced in macOS 26.1 to Maintenance → Automatic Updates
+- Add Board to System Information → Hardware → Model
+- Add CPU Identifier to System Information → Hardware → Specifications
+- Add Configuration Code to System Information → Hardware → Model
+
+### Improved
+
+- Remove “Scale to fit below built-in camera” setting from Finder Info window
+- Add “macOS” to macOS → Code Name
+- Rename “Not applicable” to “Not available”
+
+### Fixed
+
+- Fix System Information → Hardware symbol on future devices
+- Fix macOS → Code Name on future macOS versions
+- Disable app updates in recoveryOS as Sparkle does not work there
+- Fix `SettingPicker` not always displaying options vertically
+- Fix duplicate `Section`s on macOS 13
+
+### Dependencies
+
+- Update [SFSafeSymbols](https://github.com/SFSafeSymbols/SFSafeSymbols) from [6.2.0](https://github.com/SFSafeSymbols/SFSafeSymbols/releases/tag/6.2.0) to [7.0.0](https://github.com/SFSafeSymbols/SFSafeSymbols/releases/tag/7.0.0)
+- Update [Xcode](https://developer.apple.com/xcode) from [26.0.1](https://developer.apple.com/documentation/xcode-release-notes/xcode-26_0_1-release-notes) to [26.1.1](https://developer.apple.com/documentation/xcode-release-notes/xcode-26_1-release-notes)
+- Update [Sparkle](https://github.com/sparkle-project/Sparkle) from [2.8.0](https://github.com/sparkle-project/Sparkle/releases/tag/2.8.0) to [2.8.1](https://github.com/sparkle-project/Sparkle/releases/tag/2.8.1)
+
+### [Commit History](https://github.com/F1248/Genius/compare/v0.3.0...v0.4.0)
 
 ## [0.3.0](https://github.com/F1248/Genius/releases/tag/v0.3.0)
 
@@ -21,7 +58,7 @@
 - Add buttons to reset each setting to its default value
 - Add help text to `Symbol`s when showing symbol
 
-### Changed
+### Improved
 
 - Align `content` in `TabData.body` to the top
 - Use `Form`s instead of `GroupBox`es
@@ -34,7 +71,7 @@
 
 - Add app updates
 
-### Changed
+### Improved
 
 - Group settings
 
@@ -48,8 +85,8 @@
 
 ### Dependencies
 
-- Update Xcode from [26.0.0](https://developer.apple.com/news/releases/?id=09152025m) to [26.0.1](https://developer.apple.com/news/releases/?id=09222025m)
-- Update Defaults from [9.0.5](https://github.com/sindresorhus/Defaults/releases/tag/9.0.5) to [9.0.6](https://github.com/sindresorhus/Defaults/releases/tag/9.0.6)
+- Update [Xcode](https://developer.apple.com/xcode) from [26.0.0](https://developer.apple.com/documentation/xcode-release-notes/xcode-26-release-notes) to [26.0.1](https://developer.apple.com/documentation/xcode-release-notes/xcode-26_0_1-release-notes)
+- Update [Defaults](https://github.com/sindresorhus/Defaults) from [9.0.5](https://github.com/sindresorhus/Defaults/releases/tag/9.0.5) to [9.0.6](https://github.com/sindresorhus/Defaults/releases/tag/9.0.6)
 
 ### [Commit History](https://github.com/F1248/Genius/compare/v0.1.0...v0.2.0)